+++ src/components/PetsList/PetsList.jsx
@@ -1,19 +1,27 @@
 import { useSelector } from 'react-redux';
+// import { useEffect } from 'react';
+// import { getUserOperation } from 'redux/User/userOperations';
 import { getPets } from 'redux/selectors';
 import PetsItem from 'components/PetsItem/PetsItem';
 
 const PetsList = function () {
-   const pets = useSelector(getPets);
-    
+    const pets = useSelector(getPets);
+    // const dispatch = useDispatch();
+
+    // useEffect(() => {
+    //     dispatch(getUserOperation());
+    // }, [dispatch]);
+
+    console.log(pets);
 
     return (
         <>
             {pets.items.map(
-                ({ id, name, birthday, breed, comments, imgUrl }) => {
+                ({ _id, name, birthday, breed, comments, imgUrl }) => {
                     return (
                         <PetsItem
-                            key={id}
-                            id={id}
+                            key={_id}
+                            _id={_id}
                             name={name}
                             birthday={birthday}
                             breed={breed}
@@ -30,4 +38,3 @@ const PetsList = function () {
 };
 
 export default PetsList;
-
