+++ src/components/Routes/PublicRoute.js
@@ -1,13 +1,13 @@
-import { Navigate, Outlet } from "react-router-dom";
-import { useSelector } from "react-redux";
-import { getAuth } from "redux/selectors";
+import { Navigate, Outlet } from 'react-router-dom';
+import { useSelector } from 'react-redux';
+import { getAuth } from 'redux/selectors';
 
-export default function PublickRoute() {
+export default function PublicRoute() {
     const authSelector = useSelector(getAuth);
 
     if (authSelector.isLoggedIn) {
-        return <Navigate to='/notices' />
+        return <Navigate to="/user" />;
     }
-    
-    return <Outlet/>
-};
+
+    return <Outlet />;
+}
