+++ src/pages/User/UserPage.styled.jsx
@@ -1,5 +1,8 @@
 import styled from '@emotion/styled';
 
 export const Container = styled.div`
-    background: #fdf7f2;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
 `;
