+++ src/redux/User/userSlice.js
@@ -10,6 +10,7 @@ const userSlice = createSlice({
     },
     extraReducers: {
         [getUserOperation.fulfilled]: (state, { payload }) => {
+            console.log('Aaron', payload);
             return {
                 ...state,
                 user: { ...payload.user },
@@ -21,6 +22,7 @@ const userSlice = createSlice({
             return { ...state, isLoading: true };
         },
         [getUserOperation.rejected]: (state, { payload }) => {
+            console.log('AaronErr', payload);
             return { ...state, error: payload.name, isLoading: false };
         },
